requirements.txt
psycopg2
pep8
cql
cassandra-driver